requirements.txt
SQLAlchemy
uvicorn
alembic
fastapi
pydantic
pydantic-settings
python-dotenv
sqlalchemy